+++ src/GestureApp/model/Model.h
@@ -8,7 +8,7 @@ class Model
 {
 public:
 	Model(const std::string& model_dir, const int& t, const int& f) :model(model_dir), num_timestep(t), num_features(f) {};
-	int predict(const std::vector<float> &input_data);
+	int predict(const std::vector<double> &input_data);
 	int getTimeStep() const { return this->num_timestep; }
 	int getNumFeatures() const { return this->num_features; }
